Implement form requestHandler as an express router
This allow constructors to add middleware to the stack to be called before request handling using a familiar API - e.g. `use`

An example of this is in the renewal journey where the app submission needs to be checked before serving pages. It will also allow a clean-up of the photo errors model instantiation in the photo journey.