Fix postcode regex
A missing set of parentheses in the postcode regex causes the postcode test to be too loose at it allows any partial postcode match to be valid instead of insisting that the entire postcode is valid. In particular it allows a valid psotcode appended to itself to still be valid.

Wrapping the entire regex in parens - `^(...)$` forces the whole postcode to match.