closes #45 - Abstract forking logic.
Sometimes we want to extend the getNextStep method and this means there is no way to get the unaltered result of a forking conditional. A specific case is when we override getNextStep to append /edit and redirect to the /confirm page. As forking conditionals are used for step invalidation in this case the rest of the form is invalidated rather than just skiping to the confirm step.

* Added new _getForkTarget method which executes forking logic. This is called in getNextStep so the result of just calling controller.getNextStep will be the same as before
* Added public method getForkTarget to be used in form-wizard