Fix to prevent an options validator being added on every step submission
This is an initial fix for issue #42. For 'options' fields the form controller adds a validator to check that the chosen option is one of the defined set of options for the field. Rather than adding this validator on each form submission, this change adds it once on start up.

- Moves the equality validator on 'options' outside of the function return
- Unit test added for the updated code, and others cleaned up slightly.