Allow external links in redirects during failed validation
When an app is mounted on a `baseUrl` and redirects when validation fails, the baseUrl is added to the `redirect` property when it isn't '/'. This shouldn't happen when redirecting to another site. Add a check for http/s when adding 'baseUrl' so that the redirect property isn't prefixed when the redirect is another site.