Example configuration for common validators
Added a more detailed description of the `exactlength` validator configuration to provide a template for using the common validators that require extra configuration. Configuration examples for both the fields and error token setup are given. There is also a slightly more in depth look at how the `before` validator could be used.