Add `configure` method to request pipeline for dynamic configuration
In a number of instances - particularly when wanting to set dynamic field options - implementations are mutating the global form configuration to set per-session specific config. This has significant risks in a highly concurrent multi-user environment, as changes to the global config from a user's session data can be exposed to other users.

As a result it is necessary to expose an interface to set dynamic form configuration on a per-request/per-session basis.

To do this, a `_configure` method is added to the beginning of the `get`/`post` middleware pipelines, which takes a deep clone of the form configuration and exposes it to `req.form.options`, it then calls through to an implementable `configure` method, where an implementation can safely extend the form configuration in a scope limited to a particular request cycle.

Then later stages of the middleware pipeline, where previously `this.options` was being accessed, will now access `req.form.options` to be able to read any dynamic configuration.