Split call to form.locals into separate assign call
The current implementation means that the values set to `res.locals` are not on `res.locals` when `this.locals` is called. However, the extension of this class has functionality that expects the values to exist on `res.locals` in order to populate dynamic titles.

By splitting the `Object.assign` into two calls we can be sure that `res.locals.values` exists when `this.locals` is called, and dynamic content can render correctly.