Remove autocomplete functionality
This was intended to allow steps which have no fields, and so do not need to do anything on POST to simply link through to subsequent steps and save a round trip.

However, this is causing other steps which *do* need to perform actions on POST, and should not autocomplete, to incorrectly mark themselves complete. In particular "check answers" pages skipping directly to confirmation pages without submitting in some edge cases.

Since this behaviour is not known to be currently used, and is simple to implement as an extended behaviour, remove it.