Allow a fork to be marked `continueOnEdit`
When we submit a forking step in an edit journey it may be the case that one option should return to /confirm and the other option should continue along the fork. For example a step that asks a "yes/no" question about adding additional addresses - "yes" should take the user around the address journey, but "no" should return the user to /confirm.

In this case allow the fork itself to have a `continueOnEdit` option, which will allow that fork to be followed in an "edit" scenario.