Allow locals to be called in a context where req.form.errors is not set
Calling locals during the POST lifecycle in some behaviours (specifically PDF generation in firearms licensing app) causes an error because `req.form.errors` is undefined - since `getErrors` is not called as part of the POST pipeline.

By checking that the property exists, and responding accordingly if it does not this error can be prevented.

Also removes an unnecessary stub from the tests, which will improve the test coverage since it tests the actual behaviour of that method now.